wait for device info in complete intake tests, similar to recent change to review page tests
