+++ spec/rails_helper.rb
@@ -96,6 +96,7 @@ RSpec.configure do |config|
   config.include NavigationHelpers
   config.include FeatureHelpers, type: :feature
   config.include ResponsiveHelper, type: :feature
+  config.include JavascriptHelpers, type: :feature
   # RSpec Rails can automatically mix in different behaviours to your tests
   # based on their file location, for example enabling you to call `get` and
   # `post` in specs under `spec/controllers`.
